fix: harden audio pipeline with streaming output and validation

+++ Source/Host.h
@@ -8,13 +8,15 @@
 
 class Host {
 public:
+    // outputFilePath: caminho completo do arquivo de saída (com extensão).
+    // previewStartTime: em segundos (aceita fração, ex.: 0.08).
     bool processAudioFile(const std::string& pluginPath,
                             const std::vector<std::pair<int, float>>& params,
                             const std::string& inputFilePath,
                             const std::string& outputFilePath,
                             bool isPreview,
                             bool fadeOut,
-                            int previewStartTime /* em segundos, default = 0 */); 
+                            float previewStartTime);
 };
 
 #endif // HOST_H
